% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_site_bins.R
\name{.add_fixed_background}
\alias{.add_fixed_background}
\title{Ensures background megabase mutation rates use legal coordinates}
\usage{
.add_fixed_background(
  site_index,
  sites_mid,
  sites_chr,
  chr_max,
  background_width
)
}
\arguments{
\item{site_index}{Numeric index of site}

\item{sites_mid}{Numeric vector of site midpoints}

\item{sites_chr}{Character vector of site chromosomes}

\item{chr_max}{Numeric vector of max coordinate per chromosome}

\item{background_width}{Integer indicating background size (1e6)}
}
\value{
Data frame of legal site chromosomes and coordinates
}
\description{
Ensures background megabase mutation rates use legal coordinates
}
