% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_trinucleotides.R
\name{.get_all_trinucs}
\alias{.get_all_trinucs}
\title{Return possible pyrimidine-centered trinucleotide contexts}
\usage{
.get_all_trinucs()
}
\value{
Character vector of all possible trinucleotide mutation contexts with pyrimidines in the middle position
}
\description{
Return possible pyrimidine-centered trinucleotide contexts
}
