% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_flanks_nucleotides.R
\name{.get_down_flank}
\alias{.get_down_flank}
\title{Determines downstream coordinates of flanking regions to sites}
\usage{
.get_down_flank(gr_sites)
}
\arguments{
\item{gr_sites}{GenomicRanges object of sites}
}
\value{
GenomicRanges object containing the coordinates of downstream flanking regions
}
\description{
Determines downstream coordinates of flanking regions to sites
}
