% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_final_pretest_dataframe.R
\name{.get_muts}
\alias{.get_muts}
\title{Counts number of mutations per quadnucleotide context}
\usage{
.get_muts(gr_sites, gr_maf)
}
\arguments{
\item{gr_sites}{GenomicRanges object of site/flank chromosomes and positions}

\item{gr_maf}{GenomicRanges object with mcols containing quadnucleotide context or indel}
}
\value{
Data frame containing mutation count and each quadnucleotide context + indel
}
\description{
Counts number of mutations per quadnucleotide context
}
