% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_flanks_nucleotides.R
\name{.get_nucs}
\alias{.get_nucs}
\title{Creates a data frame of quadnucleotide mutation contexts with counts of corresponding trinucleotides}
\usage{
.get_nucs(gr_sites, site_ids)
}
\arguments{
\item{gr_sites}{GenomicRanges object of sites}

\item{site_ids}{Character vector of unique site ids}
}
\value{
Data frame of all possible quadnucleotide mutation contexts with the number of positions at which the trinucleotides occur
}
\description{
Creates a data frame of quadnucleotide mutation contexts with counts of corresponding trinucleotides
}
