% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_trinucleotides.R
\name{.get_quadnuc_map}
\alias{.get_quadnuc_map}
\title{Determines all trinucleotide, pyrimidine-centered trinucleotide contexts and mutation/quadnucleotide contexts}
\usage{
.get_quadnuc_map()
}
\value{
Data frame trinucleotides, alternate allele and possible quadnucleotide contexts
}
\description{
Determines all trinucleotide, pyrimidine-centered trinucleotide contexts and mutation/quadnucleotide contexts
}
