% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_trinucleotides.R
\name{.get_sequence_trinucleotides}
\alias{.get_sequence_trinucleotides}
\title{Counts the number of each trinucleotide within each site}
\usage{
.get_sequence_trinucleotides(gr_sites, site_ids)
}
\arguments{
\item{gr_sites}{GenomicRanges object of sites}

\item{side_ids}{Character vector of site ids}
}
\value{
Data frame containing each site id (character) and its trinucleotide counts (numeric)
}
\description{
Counts the number of each trinucleotide within each site
}
