% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_trinucleotides.R
\name{.get_trinucs2}
\alias{.get_trinucs2}
\title{Counts the number of each trinucleotide within each site}
\usage{
.get_trinucs2(gr_seq)
}
\arguments{
\item{gr_seq}{GenomicRanges object of sites}
}
\value{
Data frame containing each site id (character) and its trinucleotide counts (numeric)
}
\description{
Counts the number of each trinucleotide within each site
}
