% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_flanks_nucleotides.R
\name{.get_up_flank}
\alias{.get_up_flank}
\title{Determines upstream coordinates of flanking regions to sites}
\usage{
.get_up_flank(gr_sites)
}
\arguments{
\item{gr_sites}{GenomicRanges object of sites}
}
\value{
GenomicRanges object containing the coordinates of upstream flanking regions
}
\description{
Determines upstream coordinates of flanking regions to sites
}
