% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_final_pretest_dataframe.R
\name{.maf_to_dfr2}
\alias{.maf_to_dfr2}
\title{Combines sites and mutations for sites and flanks from each bin in prepared_sites}
\usage{
.maf_to_dfr2(this_maf, prepared_sites)
}
\arguments{
\item{this_maf}{Data frame of mutations prepared by get_mut_trinuc_strand. Contains all if not cofactors are included}

\item{prepared_sites}{List of coordinates and quadnucleotide contexts generated by prepare_bins_of_sites}
}
\value{
Data frame containing mutation counts for each quadnucleotide context + indel from each bin and sites + up/down flanks
}
\description{
Combines sites and mutations for sites and flanks from each bin in prepared_sites
}
