% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_final_pretest_dataframe.R
\name{.merge_sites_and_muts}
\alias{.merge_sites_and_muts}
\title{Combines quadnucleotide counts and site information}
\usage{
.merge_sites_and_muts(gr_seq, gr_maf, trinuc, is_site)
}
\arguments{
\item{gr_seq}{GenomicRanges object of site/flanks chromosomes and positions}

\item{gr_maf}{GenomicRanges object with mcols containing quadnucleotide context or indel}

\item{trinuc}{Data frame of counts for each trinucleotide counts extended to each quadnucleotide}

\item{is_site}{Boolean indicating whether mutation is within site or flanking region}
}
\value{
Data frame containing mutation and trinucleotide counts for each quadnucleotide context + indel. Includes boolean indicator of is_site
}
\description{
Combines quadnucleotide counts and site information
}
