% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_glm.R
\name{.predict_muts_ranges_probs}
\alias{.predict_muts_ranges_probs}
\title{Extract expected mutations count estimates by probability sampling}
\usage{
.predict_muts_ranges_probs(hyp, select_positions)
}
\arguments{
\item{hyp}{List containing glm.nb regression outputs}

\item{select_positions}{Numeric vector containing indices of site positions}
}
\value{
Vector of mid, lo and hi mutation count predictions
}
\description{
Extract expected mutations count estimates by probability sampling
}
