% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_site_bins.R
\name{.prepare_bins_of_sites}
\alias{.prepare_bins_of_sites}
\title{Generates bins and calls prepare_sites such that site coordinates, number of trinucleotide contexts and indels, and all possible quadnucleotide contexts are returned for each mutation rate bin and site + flank}
\usage{
.prepare_bins_of_sites(
  sites,
  window_size,
  maf,
  n_bin = NA,
  n_min_mut = 100,
  global_mut_rate_window = 1e+06
)
}
\arguments{
\item{sites}{Data frame of sites}

\item{window_size}{Integer indicating the half-width of sites and flanking regions}

\item{maf}{Data frame of mutations prepared by get_mut_trinuc_strand}

\item{n_bin}{Integer (10) indicating the number of bins to which to split sites}

\item{n_min_mut}{Integer (100) indicating minimum number of mutations to proceed with analysis}

\item{global_mut_rate_window}{Integer (1e6) indicating the window for mutation rate binning}
}
\value{
List of length n_bin each containing 3 GRanges of coordinates (site and flanks) and 3 dataframes of trinucleotide counts and mutation contexts
}
\description{
Generates bins and calls prepare_sites such that site coordinates, number of trinucleotide contexts and indels, and all possible quadnucleotide contexts are returned for each mutation rate bin and site + flank
}
