% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_site_bins.R
\name{.prepare_sites}
\alias{.prepare_sites}
\title{For each mutation rate bin and site + flanks, calculates site coordinates, number of trinucleotide contexts and indels, and all possible mutation contexts}
\usage{
.prepare_sites(sites, window_size)
}
\arguments{
\item{sites}{Data frame of sites}

\item{window_size}{Integer indicating the half-width of sites and flanking regions}
}
\value{
List containing 3 GRanges of coordinates (site and flanks) and 3 dataframes of trinucleotide counts and mutation contexts
}
\description{
For each mutation rate bin and site + flanks, calculates site coordinates, number of trinucleotide contexts and indels, and all possible mutation contexts
}
