% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_site_bins.R
\name{.prepare_sites2}
\alias{.prepare_sites2}
\title{Get merged and trimmed sites and flanking regions and trinucleotide counts}
\usage{
.prepare_sites2(this_sites, window_size)
}
\arguments{
\item{this_sites}{Data frame of sites}

\item{window_size}{Integer indicating the half-width of sites and flanking regions}
}
\value{
List containing two GRanges of sites and flanks, and two dataframes for their trinucleotide counts and mutation contexts
}
\description{
Get merged and trimmed sites and flanking regions and trinucleotide counts
}
