% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_glm.R
\name{.shall_it_pass}
\alias{.shall_it_pass}
\title{Determines whether error is permitted}
\usage{
.shall_it_pass(this_err)
}
\arguments{
\item{this_err}{Character error message}
}
\value{
Boolean indicating whether the error is permitted or not
}
\description{
Determines whether error is permitted
}
