% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_glm.R
\name{.sort_coords}
\alias{.sort_coords}
\title{Sort by increasing chromosomes and positions}
\usage{
.sort_coords(dfr)
}
\arguments{
\item{dfr}{Data Frame containing columns "chr", "start" and "end"}
}
\value{
Data Frame ordered
}
\description{
Sort by increasing chromosomes and positions
}
