% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_trinucleotides.R
\name{.split_seq_trinucs}
\alias{.split_seq_trinucs}
\title{Extract trinucleotides from sequence}
\usage{
.split_seq_trinucs(x)
}
\arguments{
\item{x}{List from strsplit containing site sequences}
}
\value{
Character vector of sequence trinucleotides
}
\description{
Extract trinucleotides from sequence
}
