% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_glm.R
\name{.test_NB_model}
\alias{.test_NB_model}
\title{Runs negative binomial regression and likelihood ratio test}
\usage{
.test_NB_model(dfr, mut_type, n_min_mut = 100, test_cofactor = F)
}
\arguments{
\item{dfr}{Data frame generated from maf_to_dfr that contains mutation and trinucleotide counts for each quadnucleotide context + indel}

\item{mut_type}{Character corresponding to total mutations or name of subclass}

\item{n_min_mut}{Integer (default 100) representing the minimum number of mutations required to run test}

\item{test_cofactor}{Boolean indicating whether to analyze cofactor}
}
\value{
Data frame containing the results of the regression with the following columns: mut_type,
pp, this_coef, obs_mut, exp_mut, exp_mut_lo, exp_mut_hi, fc, pp_cofac, this_coef_cofac
\describe{
\item{mut_type}{A string identifying the mutation class}
\item{pp}{The p-value from the likelihood ratio test}
\item{this_coef}{The coefficient from is_site}
\item{obs_mut}{The total number of observed mutations of that class}
\item{exp_mut}{The expected number of mutations determined by the model}
\item{exp_mut_lo}{Lower bound of 95\% confidence interval}
\item{exp_mut_hi}{Upper bound of 95\% confidence interval}
\item{fc}{Observed mutations divided by expected mutations}
\item{pp_cofac}{The p-value from the likelihood ratio test of site:cofactor interaction}
\item{this_coef_cofac}{The coefficient from the site:cofactor interaction term}
}
}
\description{
Runs negative binomial regression and likelihood ratio test
}
