% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_site_bins.R
\name{.trim_chrom_boundaries}
\alias{.trim_chrom_boundaries}
\title{Clean sites near chromosome boundaries}
\usage{
.trim_chrom_boundaries(gr_sites_this)
}
\arguments{
\item{gr_sites_this}{GRanges object of sites}
}
\value{
GRanges object containing cleaned and trimmed sites
}
\description{
Clean sites near chromosome boundaries
}
