% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mutation_patterns.R
\name{get_mutations_in_flanked_sites}
\alias{get_mutations_in_flanked_sites}
\title{Mutation counts along relative positions}
\usage{
get_mutations_in_flanked_sites(mutations, this_sites, window_size, n_patients)
}
\arguments{
\item{mutations}{Data frame of mutations}

\item{this_sites}{Data frame of sites}

\item{window_size}{Integer indicating the half-width of sites and flanking regions (added to both sides of midpoint)}

\item{n_patients}{Integer indicating number of patients}
}
\value{
Data frame containing the frequency of mutations across site and flank positions
\describe{
\item{dist}{the relative position of each base to nearest site midpoint}
\item{mut_count}{the number of overlapping mutations}
\item{is_site}{boolean indicating if the position is within the site}
\item{mut_freq}{frequency of mutations calculated as mut_count / n_patients / total bp * 1e6}
}
}
\description{
Prepares data frame of mutations at relative positions along flanks and sites
}
