% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mutation_patterns.R
\name{plot_mutations_in_flanked_sites}
\alias{plot_mutations_in_flanked_sites}
\title{Visualizes mutations in sites and flanks}
\usage{
plot_mutations_in_flanked_sites(dfr, window_size = 100, y_range = NULL)
}
\arguments{
\item{dfr}{Data frame of mutation counts at relative positions along stacked sites/flanks
\describe{
\item{dist}{relative position of mutations}
\item{count}{number of mutations at relative position}
\item{is_site}{boolean indicator of whether relative position is within the site}
}}

\item{window_size}{Half of the site width. Added to each side of site midpoint}
}
\value{
ggplot object
}
\description{
Visualizes mutations in sites and flanks
}
