% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_glm.R
\name{RM2_downsample}
\alias{RM2_downsample}
\title{Evaluating differential mutation rates across classes of sites with downsampling}
\usage{
RM2_downsample(
  maf,
  sites,
  mut_class_columns = NA,
  cofactor_column = NA,
  window_size = 100,
  n_min_mut = 100,
  n_bin = 10,
  n_sites_sampled,
  n_iterations = 100
)
}
\arguments{
\item{maf}{Data frame of mutations
\describe{
\item{chr}{autosomal chromosomes as chr1 to chr22 and sex chromosomes as chrX and chrY}
\item{start}{the start position of the mutation in base 1 coordinates}
\item{end}{the end position of the mutation in base 1 coordinates}
\item{ref}{the reference allele as a string containing the bases A, T, C or G}
\item{alt}{the alternate allele as a string containing the bases A, T, C or G}
\item{mut_trinuc}{trinucleotide context - where middle is C or T - with alternate allele}
\item{mut_strand}{character indicating Watson (w) or Crick (c)}
\item{ref_alt}{character indicating single-base substitution}
}}

\item{sites}{Data frame
\describe{
\item{chr}{autosomal chromosomes as chr1 to chr22 and sex chromosomes as chrX and chrY}
\item{start}{the start position of the mutation in base 0 coordinates}
\item{end}{the end position of the mutation in base 0 coordinates}
}}

\item{mut_class_columns}{Character corresponding to the column(s) of mutation classes for grouped analysis}

\item{cofactor_column}{Character corresponding to the column of cofactors}

\item{window_size}{Integer indicating the half-width of sites and flanking regions (added to left and right for full width) (default 100)}

\item{n_min_mut}{Integer indicating the minimum number of mutations required to perform analysis (default 100)}

\item{n_bin}{Integer indicating the number of megabase bins to use (default 10)}

\item{n_sites_sampled}{Integer indicating the number of sites to sample}

\item{n_iterations}{Integer indicating how many times to repeat the sampling procedure (default 100)}
}
\value{
Data frame containing the regression estimates and likelihood ratio test output with the following columns: mut_type,
pp, this_coef, obs_mut, exp_mut, exp_mut_lo, exp_mut_hi, fc, n_sites_tested
\describe{
\item{mut_type}{A string identifying the mutation class}
\item{pp}{The p-value from the likelihood ratio test}
\item{this_coef}{The coefficient from is_site}
\item{obs_mut}{The total number of observed mutations of that class}
\item{exp_mut}{The expected number of mutations determined by the model}
\item{exp_mut_lo}{Lower bound of 95\% confidence interval}
\item{exp_mut_hi}{Upper bound of 95\% confidence interval}
\item{fc}{Observed mutations divided by expected mutations}
\item{pp_cofac}{The p-value from the likelihood ratio test of site:cofactor interaction}
\item{this_coef_cofac}{The coefficient from the site:cofactor interaction term}
\item{n_sites_tested}{The number of sites that were tested - all sites if no downsampling}
}
}
\description{
RM2_downsample() is a wrapper that first performs downsampling then calls RM2. The median index is selected by p-value for total mutations (mut_class_columns=NA) and the corresponding values are returned.
}
