% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_flanks_nucleotides.R
\name{which_median}
\alias{which_median}
\title{Selects median value with priority given to larger value}
\usage{
which_median(x)
}
\arguments{
\item{x}{Numeric vector of p-values}
}
\value{
Position of median value
}
\description{
Selects median value with priority given to larger value
}
