% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mutation_patterns.R
\name{get_mutations_in_flanked_sites}
\alias{get_mutations_in_flanked_sites}
\title{Mutation counts along relative positions}
\usage{
get_mutations_in_flanked_sites(mutations, this_sites, window_size, n_patients)
}
\arguments{
\item{mutations}{Data frame of mutations}

\item{this_sites}{Data frame of sites}

\item{window_size}{Integer indicating the half-width of sites and flanking regions (added to both sides of midpoint)}

\item{n_patients}{Integer indicating number of patients}
}
\value{
Data frame containing the relative position of each base (numeric), mut_count (numeric) of mutations,
mut_freq (numeric) calulated as mut_count / n_patients / total bp * 1e6, and site status (boolean)
}
\description{
Prepares data frame of mutations at relative positions along flanks and sites
}
