% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mutation_patterns.R
\name{get_mutations_in_flanked_sites}
\alias{get_mutations_in_flanked_sites}
\title{Counts mutations at positions along flanks and sites}
\usage{
get_mutations_in_flanked_sites(mutations, sites, window_size)
}
\arguments{
\item{mutations}{Data frame of mutations}

\item{sites}{Data frame of sites}

\item{window_size}{Half the of site width}
}
\value{
Data frame containing the relative position of each base (numeric), count (numeric) of mutations and site status (boolean)
}
\description{
Counts mutations at positions along flanks and sites
}
