% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_mutation_patterns.R
\name{plot_mutations_in_flanked_sites}
\alias{plot_mutations_in_flanked_sites}
\title{Visualize mutations in sites and flanks}
\usage{
plot_mutations_in_flanked_sites(
  dfr,
  window_size = 100,
  what_plot = "mut_freq",
  y_range = NULL
)
}
\arguments{
\item{dfr}{Data frame of mutation counts at relative positions along stacked sites/flanks (from get_mutations_in_flanked_sites)
\describe{
\item{dist}{relative position of mutations}
\item{mut_count}{number of mutations at relative position}
\item{is_site}{boolean indicating whether the relative position is within the site}
\item{mut_freq}{frequency of mutations, calculated as mut_count / n_patients / total bp * 1e6 by get_mutations_in_flanked_sites}
}}

\item{window_size}{Integer indicating the half-width of sites and flanking regions (added to both sides of midpoint)}

\item{what_plot}{Character indicating y-axis variable (default "mut_freq")}

\item{y_range}{Numeric y-axis limits (default NULL)}
}
\value{
ggplot object
}
\description{
Barplot of mutations along relative positions to site midpoints
}
